# SOA/Laba2/BarsService/src/main/java/itmo/soa/barsservice/dto/ErrorResponse.java
package itmo.soa.barsservice.dto;

import org.springframework.http.HttpStatus;

import java.time.OffsetDateTime;

public class ErrorResponse {
    private OffsetDateTime timestamp;
    private int status;
    private String error;
    private String message;
    private String path;

    public static ErrorResponse from(HttpStatus httpStatus, String message, String path) {
        ErrorResponse response = new ErrorResponse();
        response.setTimestamp(OffsetDateTime.now());
        response.setStatus(httpStatus.value());
        response.setError(httpStatus.getReasonPhrase());
        response.setMessage(message);
        response.setPath(path);
        return response;
    }

    public OffsetDateTime getTimestamp() {
        return timestamp;
    }

    public void setTimestamp(OffsetDateTime timestamp) {
        this.timestamp = timestamp;
    }

    public int getStatus() {
        return status;
    }

    public void setStatus(int status) {
        this.status = status;
    }

    public String getError() {
        return error;
    }

    public void setError(String error) {
        this.error = error;
    }

    public String getMessage() {
        return message;
    }

    public void setMessage(String message) {
        this.message = message;
    }

    public String getPath() {
        return path;
    }

    public void setPath(String path) {
        this.path = path;
    }
}