# SOA/Laba2/BarsService/src/main/java/itmo/soa/barsservice/model/Coordinates.java
package itmo.soa.barsservice.model;

import com.fasterxml.jackson.annotation.JsonIgnoreProperties;

@JsonIgnoreProperties(ignoreUnknown = true)
public class Coordinates {
    private Double x;
    private Double y;

    public Coordinates() {
    }

    public Coordinates(Double x, Double y) {
        this.x = x;
        this.y = y;
    }

    public static Coordinates copyOf(Coordinates source) {
        if (source == null) {
            return null;
        }
        return new Coordinates(source.getX(), source.getY());
    }

    public Double getX() {
        return x;
    }

    public void setX(Double x) {
        this.x = x;
    }

    public Double getY() {
        return y;
    }

    public void setY(Double y) {
        this.y = y;
    }
}