# SOA/Laba2/openapi.yaml
openapi: 3.0.3
info:
    version: "1"
    title: LabWork Management API
    description: |
        API для управления коллекцией LabWorkов с поддержкой
        CRUD операций, фильтрации, сортировки, пагинации и дополнительных операций.
        
        Система состоит из двух веб-сервисов:
        1. **LabWork Management** - основной сервис для управления коллекцией
        2. **BARS Operations** - дополнительные операции над LabWorkами
servers:
    - url: http://localhost:8085/LabworkService_war
      description: Tomcat-деплой основного сервиса для разработки
...
paths:
...
    /bars/labwork/{labworkId}/difficulty/decrease/{stepsCount}:
        post:
            tags:
                - BARS Operations
            summary: Понизить сложность лабораторной работы
            description: |
                Уменьшает значение поля `difficulty` на указанное количество шагов.
                
                Порядок сложности (с индексами):
                - VERY_EASY (0) - минимальное значение
                - EASY (1)
                - INSANE (2)
                - TERRIBLE (3)
                
                Если результат операции выходит за границы (< VERY_EASY), устанавливается VERY_EASY.
            operationId: decreaseDifficulty
            parameters:
                - name: labworkId
                  in: path
                  description: ID лабораторной работы
                  required: true
                  schema:
                      type: integer
                      format: int64
                      minimum: 1
                  example: 1
                - name: stepsCount
                  in: path
                  description: Количество шагов для понижения сложности
                  required: true
                  schema:
                      type: integer
                      format: int32
                      minimum: 1
                  example: 2
            responses:
                '200':
                    description: Сложность успешно изменена
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/LabWork'
                '400':
                    description: Неверное количество шагов (должно быть > 0)
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                '404':
                    description: Лабораторная работа не найдена
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
...
    /disciplines:
        get:
            tags:
                - LabWork Management
            summary: Получить список всех дисциплин
            operationId: getAllDisciplines
            responses:
                '200':
                    description: Успешно получен список дисциплин
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/Discipline'
        post:
            tags:
                - LabWork Management
            summary: Создать новую дисциплину
            operationId: createDiscipline
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/DisciplineInput'
            responses:
                '201':
                    description: Дисциплина успешно создана
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Discipline'
                '400':
                    description: Неверные данные
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
    /disciplines/{id}:
        get:
            tags:
                - LabWork Management
            summary: Получить дисциплину по ID
            operationId: getDiscipline
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                      type: integer
                      format: int64
                      minimum: 1
            responses:
                '200':
                    description: Успешно получена дисциплина
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Discipline'
                '404':
                    description: Дисциплина не найдена
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
        put:
            tags:
                - LabWork Management
            summary: Обновить дисциплину
            operationId: updateDiscipline
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                      type: integer
                      format: int64
                      minimum: 1
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/DisciplineInput'
            responses:
                '200':
                    description: Дисциплина успешно обновлена
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Discipline'
                '400':
                    description: Неверные данные
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                '404':
                    description: Дисциплина не найдена
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
        delete:
            tags:
                - LabWork Management
            summary: Удалить дисциплину
            operationId: deleteDiscipline
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                      type: integer
                      format: int64
                      minimum: 1
            responses:
                '204':
                    description: Дисциплина успешно удалена
                '404':
                    description: Дисциплина не найдена
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
components:
    schemas:
        LabWork:
            type: object
            description: Лабораторная работа
            required:
                - name
                - coordinates
                - creationDate
                - minimalPoint
                - discipline
            properties:
                id:
                    type: integer
                    format: int64
                    description: Уникальный идентификатор (генерируется автоматически)
                    readOnly: true
                    minimum: 1
                    example: 1
                name:
                    type: string
                    description: Название лабораторной работы
                    minLength: 1
                    example: "Алгоритмы сортировки"
                coordinates:
                    $ref: '#/components/schemas/Coordinates'
                creationDate:
                    type: string
                    format: date-time
                    description: Дата и время создания (генерируется автоматически)
                    readOnly: true
                    example: "2025-10-24T12:00:00"
                minimalPoint:
                    type: number
                    format: float
                    description: Минимальный балл за работу (должен быть больше 0)
                    exclusiveMinimum: true
                    minimum: 0
                    example: 50.5
                personalQualitiesMaximum:
                    type: integer
                    format: int32
                    description: Максимальное значение личностных качеств (должно быть больше 0)
                    exclusiveMinimum: true
                    minimum: 0
                    example: 100
                difficulty:
                    $ref: '#/components/schemas/Difficulty'
                disciplineId:
                    type: integer
                    format: int64
                    description: ID дисциплины
                    minimum: 1
                    example: 1
...
        Coordinates:
            type: object
            description: Координаты лабораторной работы
            required:
                - x
                - y
            properties:
                x:
                    type: number
                    format: double64
                    description: Координата X
                    example: 100
                y:
                    type: number
                    format: float64
                    description: Координата Y (должна быть больше -156)
                    minimum: -155.0
                    example: 50
        Discipline:
            type: object
            description: Дисциплина
            required:
                - id
                - name
            properties:
                id:
                    type: integer
                    format: int64
                    description: ID дисциплины
                    minimum: 1
                    example: 1
                name:
                    type: string
                    description: Название дисциплины
                    minLength: 1
                    example: "Алгоритмы и структуры данных"
                practiceHours:
                    type: integer
                    description: Количество часов практики
                    minimum: 0
                    example: 48
        DisciplineInput:
            type: object
            description: Входные данные для дисциплины
            required:
                - name
            properties:
                name:
                    type: string
                    description: Название дисциплины
                    minLength: 1
                    example: "Алгоритмы и структуры данных"
                practiceHours:
                    type: integer
                    description: Количество часов практики
                    minimum: 0
                    example: 48
        Difficulty:
            type: string
            description: Уровень сложности лабораторной работы
            enum:
                - VERY_EASY
                - EASY
                - INSANE
                - TERRIBLE
            example: "EASY"
...
        ErrorResponse:
            type: object
            description: Описание ошибки
            properties:
                timestamp:
                    type: string
                    format: date-time
                    description: Время возникновения ошибки
                    example: "2025-11-08T12:00:00"
                status:
                    type: integer
                    format: int32
                    description: HTTP статус код
                    example: 400
                error:
                    type: string
                    description: Описание типа ошибки
                    example: "Bad Request"
                message:
                    type: string
                    description: Детальное сообщение об ошибке
                    example: "Validation failed for field 'name': must not be empty"
                path:
                    type: string
                    description: Путь запроса, вызвавшего ошибку
                    example: "/api/labworks"